Describe and updat draw_clear_frame
